+++ app/api/__init__.py
@@ -1,3 +1,5 @@
+"""The api router."""
+
 from fastapi.routing import APIRouter
 
 from . import routes