+++ app/crud/base.py
@@ -1,6 +1,5 @@
-from typing import Any, Dict, Generic, List, Optional, Type, TypeVar, Union
+from typing import Any, Generic, List, Optional, Type, TypeVar
 
-from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import HTTPException
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
@@ -25,6 +24,13 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
     def get(self, db: Session, id: Any) -> Optional[ModelType]:
         return db.query(self.model).filter(self.model.id == id).first()
 
+    def get_or_404(self, db: Session, id: Any) -> ModelType:
+        obj = self.get(db, id=id)
+        if obj is not None:
+            return obj
+        detail = f"{self.model.__name__} with id={id} does not exist"
+        raise HTTPException(404, detail)
+
     def get_multi(
         self, db: Session, *, skip: int = 0, limit: int = 100
     ) -> List[ModelType]:
@@ -32,39 +38,34 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
 
     def create(self, db: Session, *, obj_in: CreateSchemaType) -> ModelType:
         obj_in_data = obj_in.dict()
+        if "id" in obj_in_data and self.get(db, id=obj_in_data["id"]):
+            detail = f"{self.model.__name__} with id={obj_in_data['id']} already exists"
+            raise HTTPException(400, detail)
+
         db_obj = self.model(**obj_in_data)
         db.add(db_obj)
         db.commit()
         db.refresh(db_obj)
         return db_obj
 
+    @staticmethod
     def update(
-        self,
         db: Session,
         *,
         db_obj: ModelType,
-        obj_in: Union[UpdateSchemaType, Dict[str, Any]],
+        obj_in: UpdateSchemaType,
     ) -> ModelType:
 
-        obj_data = jsonable_encoder(db_obj)
-        if isinstance(obj_in, dict):
-            update_data = obj_in
-        else:
-            update_data = obj_in.dict(exclude_unset=True)
+        obj_data = obj_in.dict(exclude_unset=True)
         for field in obj_data:
-            if field in update_data:
-                setattr(db_obj, field, update_data[field])
+            setattr(db_obj, field, obj_data[field])
+
         db.add(db_obj)
         db.commit()
         db.refresh(db_obj)
         return db_obj
 
-    def remove(self, db: Session, *, id: int) -> None:
-        obj = db.query(self.model).get(id)
-        if obj is None:
-            detail = f"{self.model.__name__} with id={id} does not exist"
-            raise HTTPException(404, detail)
-
+    def remove(self, db: Session, *, id: Any) -> None:
+        obj = self.get_or_404(db, id=id)
         db.delete(obj)
         db.commit()
-        return