+++ app/tests/test_api/test_api_health_data.py
@@ -47,20 +47,39 @@ def test_health_data_get_from_user(
     assert real_heath_datas == saved
 
 
+def test_get_health_data_by_id(
+    client: TestClient, db: Session, superuser_token_headers
+):
+    user_id = random_int()
+    hd_db = crud.health_data.create(db, obj_in=gen_health_data_create(user_id, True))
+
+    response_1 = client.get(f"/health-data/{random_int()}")
+    assert response_1.status_code == 401
+
+    response_2 = client.get(
+        f"/health-data/{random_int()}", headers=superuser_token_headers
+    )
+    assert response_2.status_code == 404
+
+    response_3 = client.get(f"/health-data/{hd_db.id}", headers=superuser_token_headers)
+    assert response_3.status_code == 200
+    hd_data = HealthData(**response_3.json())
+    assert hd_data == HealthData.from_orm(hd_db)
+
+
 def test_health_datas_get_from_all_users(
     client: TestClient, db: Session, superuser_token_headers: dict
 ):
     convs_db = crud.health_data.get_multi(db)
     convs = [HealthData.from_orm(x) for x in convs_db]
 
-    response = client.get(f"/health-data", headers=superuser_token_headers)
+    response = client.get("/health-data", headers=superuser_token_headers)
     assert response.status_code == 200
     real_heath_datas = parse_obj_as(List[HealthData], response.json())
     assert real_heath_datas == convs
 
 
 def test_health_data_delete(client: TestClient, db: Session, superuser_token_headers):
-
     health_data_in = gen_health_data_create(random_int(), True)
     health_data_db = crud.health_data.create(db, obj_in=health_data_in)